Wire UCI output to SearchResult
